DepartmentServiceImplTests done

+++ src/main/java/employeeService/employeeBook/interfaces/DepartmentsService.java
@@ -9,7 +9,7 @@ public interface DepartmentsService {
     List<Employee> getEmployeesByDep(String department);
     Employee findEmployeesMinSalaryByDep(String department);
     Employee findEmployeesMaxSalaryOfDep(String department);
-    int countSummarySalaryOfDep(String department);
+    double countSummarySalaryOfDep(String department);
     double countAverageSalaryOfDep(String department);
     void toIndexSalaryOfDep(String department, int percent);
     List<Employee> printDepartment(String department);
